+++ test/ws/message.test.ts
@@ -516,4 +516,38 @@ describe('WebSocket message', () => {
 		await wsClosed(ws)
 		app.stop()
 	})
+
+    it('handle validation error with onError', async () => {
+        const app = new Elysia()
+            .onError(() => {
+                return 'caught'
+            })
+            .ws('/ws', {
+                body: t.Object({
+                    name: t.String()
+                }),
+                message(ws, message) {
+                    return ws.send(message)
+                }
+            })
+            .listen(0)
+
+        const ws = newWebsocket(app.server!)
+
+        await wsOpen(ws)
+
+        const message = wsMessage(ws)
+
+        ws.send(JSON.stringify({
+            name: 123, // expecting a string
+        }))
+
+        const { type, data } = await message
+
+        expect(type).toBe('message')
+        expect(data).toBe('caught')
+
+        await wsClosed(ws)
+        app.stop()
+    })
 })